% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmcmc-mcmc.R
\name{afmvn_mcmc}
\alias{afmvn_mcmc}
\title{MCMC for AFMVN model}
\usage{
afmvn_mcmc(
  X,
  prior,
  init,
  warm_up = 5000,
  warm_band = 100,
  n_iter = 5000,
  alpha_repropose = 10,
  alpha_step = 0.1,
  mu_step = 0.1,
  Sigma_step = 0.1
)
}
\arguments{
\item{X}{numeric matrix (n x d)}

\item{prior}{list with m0, kappa0, Psi, nu}

\item{init}{list with alpha, mu, Sigma}

\item{warm_up, warm_band, n_iter, alpha_repropose, alpha_step, mu_step, Sigma_step}{tuning parameters}
}
\value{
list with chains and acceptance rates
}
\description{
Metropolis-within-Gibbs sampler for \code{dafmvn} likelihood with NIW prior.
}
