% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vba-afmvm.R
\name{afmvn_vba}
\alias{afmvn_vba}
\title{Variational Bayes for the AFMVN model}
\usage{
afmvn_vba(
  X,
  m_eta_init,
  s2_eta_init,
  m_q_init,
  Psi_q_init,
  nu_q_init,
  m_0,
  kappa_0,
  Psi_0,
  nu_0,
  a_0 = 0,
  s0_sq = 10,
  n_quad = 20L,
  eps_alpha = 1e-08,
  eta_control = list(maxit = 300L),
  max_iter = 150,
  tol = 0.001
)
}
\arguments{
\item{X}{Data matrix.}

\item{m_eta_init}{Initial mean of q(eta).}

\item{s2_eta_init}{Initial variance of q(eta).}

\item{m_q_init}{Initial mean vector of q(mu, Lambda).}

\item{Psi_q_init}{Initial Wishart scale matrix.}

\item{nu_q_init}{Initial Wishart degrees of freedom.}

\item{m_0}{Prior mean vector.}

\item{kappa_0}{Prior scaling parameter.}

\item{Psi_0}{Prior Wishart scale matrix.}

\item{nu_0}{Prior Wishart degrees of freedom.}

\item{a_0}{Prior mean of eta.}

\item{s0_sq}{Prior variance of eta.}

\item{n_quad}{Number of quadrature nodes.}

\item{eps_alpha}{Small value used for numerical stability.}

\item{eta_control}{Control list passed to the eta optimization.}

\item{max_iter}{Maximum number of VB iterations.}

\item{tol}{Convergence tolerance.}
}
\value{
A list containing the variational parameter estimates, iteration trace,
number of iterations, and convergence indicator.
}
\description{
Fits the AFMVN model using variational Bayes.
}
