% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_est.R
\name{alpha_est}
\alias{alpha_est}
\title{Estimation of alpha}
\usage{
alpha_est(X, by = 0.01, level = NULL)
}
\arguments{
\item{X}{A n-by-D matrix.}

\item{by}{increment of the alphas}
}
\value{
list: alpha_hat is the estimated alpha, loglik is a vector of log-likelihood for each alpha
}
\description{
Estimation of alpha
}
\examples{
alpha_est(X = X, by = 0.01)
}
