% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_est.R
\name{alpha_est2}
\alias{alpha_est2}
\title{Estimation of alpha (Method 2)}
\usage{
alpha_est2(X, mu, sigma, alpha)
}
\arguments{
\item{X}{A n-by-D matrix.}

\item{mu}{kown mean}

\item{Sigma}{known sigma}
}
\value{
list: alpha_hat is the estimated alpha, loglik is a vector of log-likelihood for each alpha
}
\description{
Estimation of alpha (Method 2)
}
\examples{
alpha_est(X = X, by = 0.01)
}
