% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fold.R
\name{alpha_fold}
\alias{alpha_fold}
\title{Alpha folded transfomration of compositional data}
\usage{
alpha_fold(X, alpha, unfold = FALSE, check_one = TRUE, renorm = FALSE)
}
\arguments{
\item{X}{A numeric vector, matrix, or data frame of positive values.
If a vector is supplied, it is treated as a single row.
If matrix or data frame are supplied, they are row stochastic.}

\item{alpha}{A numeric scalar specifying the transformation parameter.
Values close to 0 (<1e-8) adpot the log-ratio transform.}

\item{unfold}{Logical; Transform to outside of the simplex when unfold = TRUE.}

\item{check_one}{Logical; Should we check X is row stochastic?}

\item{renorm}{Logical: Should we normalize rows of X?}
}
\value{
A matrix with the same dimention of X.
}
\description{
Alpha folded transfomration of compositional data
}
\examples{
X <- c(0.1, 0.6, 0.3) 
alpha_fold(X = X , alpha = 0.7)
}
