% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fold.R
\name{alpha_fold_inverse}
\alias{alpha_fold_inverse}
\title{Inverse of alpha fold transformation}
\usage{
alpha_fold_inverse(Y, alpha)
}
\arguments{
\item{Y}{A numeric vector or matrix.
If a vector is supplied, it is treated as a single row.}

\item{alpha}{A numeric scalar specifying the transformation parameter.
Values close to 0 (<1e-8) adpot the log-ratio transform.}
}
\value{
A numeric vector or matrix.
}
\description{
Inverse of alpha fold transformation
}
\examples{
alpha_fold_inverse(Y = c(0.1, 0.2), alpha = 0.1)
}
