% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_trans.R
\name{alpha_trans}
\alias{alpha_trans}
\title{Alpha transformation of compositional data}
\usage{
alpha_trans(X, alpha, check_one = TRUE, renorm = FALSE)
}
\arguments{
\item{X}{A numeric vector, matrix, or data frame of positive values.
If a vector is supplied, it is treated as a single row.
If matrix or data frame are supplied, they are row stochastic.}

\item{alpha}{A numeric scalar specifying the transformation parameter.
Values close to 0 (<1e-8) adpot the log-ratio transform.}

\item{check_one}{Logical; Should we check X is row stochastic?}

\item{renorm}{Logical: Shoudl we normalize rows of X?}
}
\value{
A matrix with the same dimention of X.
}
\description{
Alpha transformation of compositional data
}
\examples{
X <- c(0.1, 0.6, 0.3) 
alpha_trans(X = X , alpha = 0.7)
}
