% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_trans.R
\name{alpha_trans_inverse}
\alias{alpha_trans_inverse}
\title{The inverse of alpha_trans}
\usage{
alpha_trans_inverse(W, alpha)
}
\arguments{
\item{W}{A numeric vector or matrix.
If a vector is supplied, it is treated as a single row.
The row sums should be 0.}

\item{alpha}{A numeric scalar specifying the transformation parameter.
Values close to 0 (<1e-8) adpot the log-ratio transform.}
}
\value{
A numeric vector or matrix, which is the inverse of W given alpha.
}
\description{
The inverse of alpha_trans
}
\examples{
alpha_trans_inverse(alpha_trans(X = c(0.1, 0.6, 0.3), alpha = 0.7), alpha = 0.7)
}
