% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaFoldMVN-package.R
\docType{package}
\name{AlphaFoldMVN-package}
\alias{AlphaFoldMVN}
\alias{AlphaFoldMVN-package}
\title{AlphaFoldMVN: What the Package Does (One Line, Title Case)}
\description{
Provides functions for defining, simulating, and estimating the parameters of the alpha-folded multivariate normal (MVN) distribution. The package includes density, random generation, and cumulative distribution functions, along with parameter estimation routines based on maximum likelihood and method-of-moments approaches. Designed for research on non-Gaussian modeling and folded or skewed multivariate data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Xiaoke-Q/AlphaFoldMVN}
  \item Report bugs at \url{https://github.com/Xiaoke-Q/AlphaFoldMVN/issues}
}

}
\author{
\strong{Maintainer}: Xiaoke Qin \email{Xiaoke.Qin@umanitoba.ca}

}
\keyword{internal}
