% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmvn.R
\name{dafmvn}
\alias{dafmvn}
\title{Density function of alpha-folding multivariate normal distribution.}
\usage{
dafmvn(X, alpha, mean, sigma, log = TRUE)
}
\arguments{
\item{X}{A n-by-D matrix.}

\item{alpha}{A numeric parameter of alpha-folding multivariate normal distribution.}

\item{mean}{A vector. The mean of latent multivariate normal distribution.}

\item{sigma}{A symmetric psitive-definite matrix.}

\item{log}{Logical; return log-likelihood function when log = TRUE.}
}
\value{
A vector of density values for every observation.
}
\description{
Density function of alpha-folding multivariate normal distribution.
}
\examples{
dafmvn(X = X, alpha = 0.7, mean = rep(1, D-1), sigma = diag(1, D-1))
}
