% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert.R
\name{helmert}
\alias{helmert}
\title{Generate an orthonormal Helmertian matrix in the strict sense, without the first row.}
\usage{
helmert(d)
}
\arguments{
\item{d}{The number of columns of the matrix.}
}
\value{
A (d-1)-by-d Helmertian matrix.
}
\description{
Generate an orthonormal Helmertian matrix in the strict sense, without the first row.
}
\examples{
helmert(d = 4)
}
