% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize the data points of length 3 and boundaries.}
\usage{
visualize(
  W = NULL,
  alpha,
  group = NULL,
  shift = NULL,
  color = NULL,
  opacity = NULL
)
}
\arguments{
\item{W}{A n-by-3 matrix.}

\item{alpha}{A numeric parameter of alpha transformation.}

\item{group}{A vector of length n, labeling the groups.}

\item{shift}{A vector of length n, labeling the shifts.}

\item{color}{A vector of length n, labeling the colors.}

\item{opacity}{A vector of length n, labeling the transparency.}
}
\value{
A plotly object.
}
\description{
Visualize the data points of length 3 and boundaries.
}
\examples{
visualize(W= W, alpha = 0.7, group = rep(1:3, each = nrow(W)))
}
