% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_x}
\alias{visualize_x}
\title{Visualize observed data X}
\usage{
visualize_x(X, group = NULL, color = NULL)
}
\arguments{
\item{X}{A matrix}

\item{group}{Group of data}

\item{color}{Color}
}
\value{
A plotly object
}
\description{
Visualize observed data X
}
\examples{
visualize_x(X = X)
}
