% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afmvn.R
\name{afmvn}
\alias{afmvn}
\alias{rafmvn}
\title{The alpha-folding multivariate normal distribution}
\usage{
rafmvn(n = 1, alpha, mean, sigma)
}
\arguments{
\item{n}{A scalar. Number of observations.}

\item{alpha}{A numeric parameter of the \eqn{\alpha}-transformation applied to compositional data.}

\item{mean}{A vector. The mean of latent multivariate normal distribution.}

\item{sigma}{A symmetric psitive-definite matrix.}
}
\value{
A list list(X, folded).
}
\description{
The alpha-folding multivariate normal distribution
}
\examples{
rafmvn(n = 50, alpha = 0.7, mean = c(-1,1), sigma = diag(1, 2))
}
