% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{afmvn_mle}
\alias{afmvn_mle}
\title{MLE of alpha-folding multivariate normal distribution given alpha}
\usage{
afmvn_mle(X, alpha, tol = 1e-06, max.iter = 150)
}
\arguments{
\item{X}{A n-by-D matrix.}

\item{alpha}{A numeric parameter of alpha-folding multivariate normal distribution.}

\item{tol}{A threshold of alpha. When |alpha| < tol, the clt is adopted.}

\item{max.iter}{The number of alternating iterations of solving for mu and sigma.}
}
\value{
A list of estimated mu and sigma.
}
\description{
MLE of alpha-folding multivariate normal distribution given alpha
}
\examples{
afmvn_mle(X = X, alpha = 0.7, tol = 1e-6, max.iter = 150)
}
